\name{Rcppglog-package}
\alias{Rcppglog-package}
\alias{Rcppglog}
\docType{package}
\title{\packageTitle{Rcppglog}}
\description{\packageDescription{Rcppglog}}
\section{Package Content}{\packageIndices{Rcppglog}}
\author{\packageAuthor{Rcppglog}}
\section{Maintainer}{\packageMaintainer{Rcppglog}}
\keyword{package}
